+++ templates/template.go
@@ -1,0 +1,14 @@
+package main
+
+import (
+    "fmt"
+    "os"
+)
+
+func run(s string) int {
+    // Your code goes here
+}
+
+func main() {
+    fmt.Println(run(os.Args[1]))
+}
