+++ templates/template.js
@@ -1,0 +1,7 @@
+/**
+ * @param {{string}} s puzzle input in string format
+ * @returns solution flag
+ */
+run = s => {
+	// Your code goes here
+};
