+++ templates/template.py
@@ -1,0 +1,9 @@
+from runners.python import Submission
+
+class {class_name}(Submission):
+
+	def run(self, s):
+		# :param s: input in string format
+		# :return: solution flag
+		# Your code goes here
+		pass
